Extract command executor utility to utils package
